requirements.txt
asgiref==3.4.1
Babel==2.10.3
Django==3.2.13
django-annoying==0.10.6
django-localflavor==3.1
django-phonenumber-field==6.0.0
phonenumbers==8.12.51
psycopg2-binary==2.9.3
python-stdnum==1.17
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
pytest~=7.0.1